pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["my <jakha921@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.96.1"}
sqlalchemy = "^2.0.16"
alembic = "^1.11.1"
psycopg2 = "^2.9.6"
asyncpg = "^0.27.0"
gunicorn = "^20.1.0"
flower = "^1.2.0"
pyjwt = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
